worker/requirements.txt
prometheus-client
azure-eventhub
